EndPoint para Notafiscal, Pedido e ItemPedido

+++ api/Models/Salgado.cs
@@ -7,7 +7,7 @@ namespace Art_Sabores.Models
         [Key]
         public int Id { get; set; }
         public string? Nome { get; set; } 
-        public float Preco { get; set; }
+        public decimal Preco { get; set; }
         public int Estoque { get; set; }
     }
 }
